updated folder structure, modified endpoints
