+++ .gitignore
@@ -1,1 +1,2 @@
+# dependencies
 /node_modules
