+++ server.js
@@ -3,6 +3,11 @@ const bcrypt = require('bcrypt-nodejs');
 const cors = require('cors');
 const knex = require('knex');
 
+const register = require('./controllers/register');
+const signin = require('./controllers/signin');
+const profile = require('./controllers/profile');
+const image = require('./controllers/image');
+
 const db = knex({
     client: 'pg',
     connection: {
@@ -12,10 +17,6 @@ const db = knex({
         database: 'smart-brain'
     }
 });
-
-// db.select('*').from('users').then(data => {
-//     console.log(data);
-// });
 
 const app = express();
 
@@ -30,81 +31,21 @@ app.get('/', (req, res) => {
     res.send("success");
 });
 
-// -------- /signin -----------
-app.post('/signin', (req, res) => {
-    db.select('email', 'hash').from('login')
-        .where('email', '=', req.body.email)
-        .then(data => {
-            const isValid = bcrypt.compareSync(req.body.password, data[0].hash);
-            if (isValid) {
-                return db.select('*').from('users')
-                    .where('email', '=', req.body.email)
-                    .then(user => {
-                        res.json(user[0])
-                    })
-                    .catch(err => res.status(400).json('unable to get user'))
-            } else {
-                res.status(400).json('wrong credentials')
-            }
-        })
-        .catch(err => res.status(400).json('wrong credentials'))
-})
+// Using dependency injection to inject what 
+// handleSignin, handleRegister, handleProfileGet, handleImage needs
 
-// -------- /register -----------
-app.post('/register', (req, res) => {
-    const { email, name, password } = req.body;
-    const hash = bcrypt.hashSync(password);
-        db.transaction(trx => {
-            trx.insert({
-                hash: hash,
-                email: email
-            })
-            .into('login')
-            .returning('email')
-            .then(loginEmail => {
-                return trx('users')
-                    .returning('*')
-                    .insert({
-                        email: loginEmail[0].email,
-                        name: name,
-                        joined: new Date()
-                    })
-                    .then(user => {
-                        res.json(user[0]);
-                    })
-            })
-            .then(trx.commit)
-            .catch(trx.rollback)
-        })
-        .catch(err => res.status(400).json('unable to register'))
-})
-
+// -------- /signin ----------------
+app.post('/signin', (req, res) => { signin.handleSignin(req, res, db, bcrypt) })
+// -------- /register --------------
+app.post('/register', (req, res) => { register.handleRegister(req, res, db, bcrypt) })
 // -------- /profile/:id -----------
-app.get('/profile/:id', (req, res) => {
-    const { id } = req.params;
-    db.select('*').from('users').where({id}) // id: id (ES6)
-        .then(user => {
-            if (user.length) {
-                res.json(user[0]);
-            } else {
-                res.status(400).json('Not found')
-            }
-            
-    })
-        .catch(err => res.status(400).json('error getting user'))    
-})
+app.get('/profile/:id', (req, res) => { profile.handleProfileGet(req, res, db) })
+// -------- /image -----------------
+app.put('/image', (req, res) => { image.handleImage(req, res, db) })
 
-// -------- /image -----------
-app.put('/image', (req, res) => {
-    const { id } = req.body;
-    db('users').where('id', '=', id)
-    .increment('entries', 1)
-    .returning('entries')
-    .then(entries => {
-        res.json(entries[0].entries);
-    })
-    .catch(err => res.status(400).json('unable to get entries'))
-})
+app.listen(3001, ()=> {
+    console.log('app is running on port 3001');
+});
 
 // bcrypt.hash("bacon", null, null, function(err, hash) {
 //     // Store hash in your password DB.
@@ -118,10 +59,6 @@ app.put('/image', (req, res) => {
 //     // res = false
 // });
 
-
-app.listen(3001, ()=> {
-    console.log('app is running on port 3001');
-});
 
 /*
 / --> res = this is working
@@ -141,4 +78,8 @@ CREATE TABLE users (
 	email text UNIQUE NOT NULL,
 	entries BIGINT DEFAULT 0
 );
+
+// db.select('*').from('users').then(data => {
+//     console.log(data);
+// });
 */
